fix more compatibility
